Rename occurrences package.
